refactor: 클래스 명 변경 및 패키지 변경
