+++ src/main/java/step1/CalculateMain.java
@@ -1,5 +1,6 @@
-package step1.controller;
+package step1;
 
+import step1.controller.Calculator;
 import step1.domain.Inputs;
 import step1.view.InputView;
 import step1.view.ResultView;
@@ -7,6 +8,6 @@ import step1.view.ResultView;
 public class CalculateMain {
     public static void main(String[] args) {
         Inputs inputs = InputView.start();
-        ResultView.printResult(new CalculateProcessor().calculate(inputs));
+        ResultView.printResult(new Calculator().calculate(inputs));
     }
 }
