+++ src/main/java/step1/controller/Calculator.java
@@ -3,7 +3,7 @@ package step1.controller;
 import step1.domain.Number;
 import step1.domain.Inputs;
 
-public class CalculateProcessor {
+public class Calculator {
     public Number calculate(Inputs inputs) {
         return inputs.sum();
     }
